Incorporate CadGenerator (#48)
### Description of the Change

This PR incorporates `CadGenerator` so it can be used to generate robot CAD along with a `KinematicBase`. This PR also binds a `GitScriptFactory` when the user logs in to allow for various things, including scripts, to use the GitFS instance MWC has to make scripts. This PR also includes numerous other bug fixes.